+++ R/deprecate-args.R
@@ -15,16 +15,14 @@ deprecate_arg <- function(
   old_name,
   new_name,
   fn_name,
-  version = "0.5.0"
+  version = "0.7.0"
 ) {
   if (lifecycle::is_present(old_arg)) {
-    lifecycle::deprecate_warn(
+    lifecycle::deprecate_stop(
       version,
       paste0(fn_name, "(", old_name, ")"),
       paste0(fn_name, "(", new_name, ")")
     )
-    old_arg
-  } else {
-    new_arg
   }
+  new_arg
 }