+++ R/deprecated.R
@@ -40,8 +40,8 @@
 #' @return a survey in the correct format
 #' @export
 get_survey <- function(survey, clear_cache = FALSE, ...) {
-  lifecycle::deprecate_warn(
-    when = "0.5.0",
+  lifecycle::deprecate_stop(
+    when = "0.7.0",
     what = "get_survey()",
     with = "contactsurveys::download_survey()",
     details = "We recommend using contactsurveys::download_survey() to \\
@@ -117,8 +117,8 @@ get_survey <- function(survey, clear_cache = FALSE, ...) {
 #' @export
 # nolint start: cyclocomp_linter.
 download_survey <- function(survey, dir = NULL, sleep = 1) {
-  lifecycle::deprecate_warn(
-    when = "0.5.0",
+  lifecycle::deprecate_stop(
+    when = "0.7.0",
     what = "download_survey()",
     with = "contactsurveys::download_survey()"
   )
@@ -254,8 +254,8 @@ download_survey <- function(survey, dir = NULL, sleep = 1) {
 #' }
 #' @export
 list_surveys <- function(clear_cache = FALSE) {
-  lifecycle::deprecate_warn(
-    when = "0.5.0",
+  lifecycle::deprecate_stop(
+    when = "0.7.0",
     what = "list_surveys()",
     with = "contactsurveys::list_surveys()"
   )
@@ -337,8 +337,10 @@ list_surveys <- function(clear_cache = FALSE) {
 #' @return list of countries
 #' @inheritParams get_survey
 #' @examples
+#' \dontrun{
 #' data(polymod)
 #' survey_countries(polymod)
+#' }
 #' ## --> we now recommend
 #' \dontrun{
 #' doi_peru <- "10.5281/zenodo.1095664" # nolint
@@ -351,8 +353,8 @@ list_surveys <- function(clear_cache = FALSE) {
 #' }
 #' @export
 survey_countries <- function(survey, country.column = "country", ...) {
-  lifecycle::deprecate_warn(
-    when = "0.5.0",
+  lifecycle::deprecate_stop(
+    when = "0.7.0",
     what = "survey_countries()",
     with = "contactsurveys::download_survey()",
     details = "We recommend using contactsurveys::download_survey() to \\
@@ -390,8 +392,8 @@ survey_countries <- function(survey, country.column = "country", ...) {
 #' }
 #' @export
 get_citation <- function(x) {
-  lifecycle::deprecate_warn(
-    when = "0.5.0",
+  lifecycle::deprecate_stop(
+    when = "0.7.0",
     what = "get_citation()",
     with = "contactsurveys::get_citation()"
   )