+++ R/lists.R
@@ -23,8 +23,8 @@
 #' }
 #' @export
 wpp_countries <- function() {
-  lifecycle::deprecate_soft(
-    "0.6.0",
+  lifecycle::deprecate_warn(
+    "0.7.0",
     "wpp_countries()",
     details = c(
       "Pass population data directly via the \\